Fin du projet

+++ DTO/Payment.cs
@@ -10,7 +10,7 @@ namespace DTO
     {
         public int PaymentId { get; set; }
 
-        public int Quota { get; set; }
+        public decimal Quota { get; set; }
 
         public int UserId { get; set; }
